% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexa.R
\name{complexa}
\alias{complexa}
\title{Função Complexa}
\usage{
complexa(a, b)
}
\arguments{
\item{a}{primeiro argumento}

\item{b}{segundo argumento}
}
\value{
uma tibble
}
\description{
Esta função faz a divisão de uma multiplicação por uma soma
}
\examples{

}
